refactor: split auth into keyring management + identity management
Keyring commands (signing key):
- auth generate — generate new key + store in keyring (simplified, no Safe/Timelock)
- auth login — import existing key + store + discover identities on-chain
- auth logout — remove key from keyring + clear all identities

Identity commands (new):
- auth identity new — create Safe or Timelock (moved from auth generate)
- auth identity list — show all stored identities
- auth identity select — switch active identity per environment

Changes:
- auth generate no longer offers Safe/Timelock creation
- auth login no longer has identity selector (use auth identity select)
- auth logout now clears identities (prevents orphaned state)
- Added removeIdentity() to globalConfig.ts
- Added auth-flow-map.md with complete state transition table,
  decision trees for every command, and command tree with key requirements
