+++ packages/cli/src/commands/auth/login.ts
@@ -1,7 +1,8 @@
 /**
  * Auth Login Command
  *
- * Store an existing private key in OS keyring
+ * Import an existing private key into OS keyring.
+ * Automatically discovers associated Timelocks and Safes on-chain.
  */
 
 import { Command } from "@oclif/core";
@@ -19,7 +20,7 @@ import {
   discoverTimelockForEOA,
   type LegacyKey,
 } from "@layr-labs/ecloud-sdk";
-import { getHiddenInput, displayWarning, showPrivateKey } from "../../utils/security";
+import { getHiddenInput, displayWarning } from "../../utils/security";
 import { withTelemetry } from "../../telemetry";
 import { commonFlags } from "../../flags";
 import {
@@ -27,14 +28,12 @@ import {
   addIdentity,
   replaceAllIdentities,
   setActiveIdentity,
-  getActiveIdentityAddress,
-  formatIdentity,
 } from "../../utils/globalConfig";
 import { createPublicClientOnly } from "../../utils/viemClients";
 import type { Address } from "viem";
 
 export default class AuthLogin extends Command {
-  static description = "Store your private key in OS keyring, or switch active identity";
+  static description = "Import an existing private key into OS keyring";
 
   static examples = ["<%= config.bin %> <%= command.id %>"];
 
@@ -48,62 +47,20 @@ export default class AuthLogin extends Command {
       const { flags } = await this.parse(AuthLogin);
       const environment = flags.environment as string;
 
-      const identities = getIdentities();
-
-      // One or more identities — show selector
-      if (identities.length > 0) {
-        const activeAddress = getActiveIdentityAddress(environment);
-        const choices = [
-          ...identities.map((id) => ({
-            name: formatIdentity(id) + (id.address.toLowerCase() === activeAddress?.toLowerCase() ? "  ✓ active" : ""),
-            value: id.address,
-          })),
-          { name: "─── Replace signing key ───", value: "__key__" },
-        ];
-
-        const selected = await select({
-          message: `Select active identity for ${environment}:`,
-          choices,
-        });
-
-        if (selected !== "__key__") {
-          setActiveIdentity(environment, selected);
-          const id = identities.find((i) => i.address.toLowerCase() === selected.toLowerCase())!;
-          this.log(`\n✓ Active identity: ${formatIdentity(id)}`);
-          return;
-        }
-
-        // Show existing key for backup before replacing
+      // Check for existing key
+      const exists = await keyExists();
+      if (exists) {
         const existing = await getPrivateKeyWithSource({ privateKey: undefined });
         if (existing) {
           const existingAddress = getAddressFromPrivateKey(existing.key);
-          const backupContent = `
-━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
-Your existing signing key is shown below.
-Back it up before it is replaced.
-
-Address:     ${existingAddress}
-Private key: ${existing.key}
-━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
-
-Press 'q' to exit and continue...
-`;
-          await showPrivateKey(backupContent);
+          displayWarning([
+            "A signing key already exists.",
+            `Address: ${existingAddress}`,
+            "",
+            "Replacing it will clear all current identities.",
+          ]);
         }
-
-        // User chose to replace signing key — warn before proceeding
-        displayWarning([
-          "A signing key already exists.",
-          "Replacing it will clear all current identities.",
-          "Make sure you have backed up your existing key.",
-        ]);
-        this.log("");
-
-        const confirmReplace = await confirm({
-          message: "Replace current signing key?",
-          default: false,
-        });
-
+        const confirmReplace = await confirm({ message: "Replace current signing key?", default: false });
         if (!confirmReplace) {
           this.log("\nCancelled.");
           return;
@@ -119,7 +76,6 @@ Press 'q' to exit and continue...
         this.log("\nFound legacy keys from eigenx-cli:");
         this.log("");
 
-        // Display legacy keys
         for (const key of legacyKeys) {
           this.log(`  Address: ${key.address}`);
           this.log(`  Environment: ${key.environment}`);
@@ -133,7 +89,6 @@ Press 'q' to exit and continue...
         });
 
         if (importLegacy) {
-          // Create choices for selection
           const choices = legacyKeys.map((key) => ({
             name: `${key.address} (${key.environment} - ${key.source})`,
             value: key,
@@ -144,7 +99,6 @@ Press 'q' to exit and continue...
             choices,
           });
 
-          // Retrieve the actual private key
           privateKey = await getLegacyPrivateKey(selectedKey.environment, selectedKey.source);
 
           if (!privateKey) {
@@ -158,7 +112,6 @@ Press 'q' to exit and continue...
       // If no legacy key was selected, prompt for private key input
       if (!privateKey) {
         privateKey = await getHiddenInput("Enter your private key:");
-
         privateKey = privateKey.trim();
       }
 
@@ -166,9 +119,7 @@ Press 'q' to exit and continue...
         this.error("Invalid private key format. Please check and try again.");
       }
 
-      // Derive address for confirmation
       const address = getAddressFromPrivateKey(privateKey);
-
       this.log(`\nAddress: ${address}`);
 
       const confirmStore = await confirm({
@@ -181,19 +132,17 @@ Press 'q' to exit and continue...
         return;
       }
 
-      // Store in keyring
       try {
         await storePrivateKey(privateKey);
         this.log("\n✓ Private key stored in OS keyring");
         this.log(`✓ Address: ${address}`);
-        this.log("You can now use ecloud commands without --private-key flag.");
 
-        // Switching signing key — wipe all identities (they belonged to the previous EOA)
+        // New signing key — wipe old identities, set EOA as default
         replaceAllIdentities([{ type: "eoa", address }]);
         setActiveIdentity(environment, address);
 
-        // Discover canonical Timelock on-chain for this EOA
-        this.log(`\nScanning chain for Timelock associated with ${address}...`);
+        // Discover Timelock on-chain
+        this.log(`\nScanning chain for associated identities...`);
         try {
           const publicClient = createPublicClientOnly({ environment, rpcUrl: flags["rpc-url"] });
           const environmentConfig = getEnvironmentConfig(environment);
@@ -223,7 +172,7 @@ Press 'q' to exit and continue...
           this.log(`(Timelock scan skipped — chain not reachable)`);
         }
 
-        // Discover Safes where this EOA is an owner via Safe Transaction Service
+        // Discover Safes
         try {
           const safeServiceUrl =
             environment === "mainnet-alpha"
@@ -255,7 +204,7 @@ Press 'q' to exit and continue...
           // Safe Transaction Service not reachable — skip
         }
 
-        // Ask if user wants to delete the legacy key (only if save was successful)
+        // Clean up legacy key if imported
         if (selectedKey) {
           this.log("");
           const confirmDelete = await confirm({
@@ -264,28 +213,17 @@ Press 'q' to exit and continue...
           });
 
           if (confirmDelete) {
-            const deleted = await deleteLegacyPrivateKey(
-              selectedKey.environment,
-              selectedKey.source,
-            );
-
+            const deleted = await deleteLegacyPrivateKey(selectedKey.environment, selectedKey.source);
             if (deleted) {
-              this.log(
-                `\n✓ Legacy key deleted from ${selectedKey.source}:${selectedKey.environment}`,
-              );
-              this.log("\nNote: The key is now only stored in ecloud. You can still use it with");
-              this.log("eigenx-cli by providing --private-key flag or EIGENX_PRIVATE_KEY env var.");
+              this.log(`\n✓ Legacy key deleted from ${selectedKey.source}:${selectedKey.environment}`);
             } else {
-              this.log(
-                `\n⚠️  Failed to delete legacy key from ${selectedKey.source}:${selectedKey.environment}`,
-              );
-              this.log("The key may have already been removed.");
+              this.log(`\n⚠️  Failed to delete legacy key`);
             }
-          } else {
-            this.log(`\nLegacy key kept in ${selectedKey.source}:${selectedKey.environment}`);
-            this.log("You can delete it later using 'eigenx auth logout' if needed.");
           }
         }
+
+        this.log("\nRun 'ecloud auth identity new' to create a Safe or Timelock identity.");
+        this.log("Run 'ecloud auth identity select' to switch active identity.");
       } catch (err: any) {
         this.error(`Failed to store key: ${err.message}`);
       }
